refactor: move reader function to io module
