+++ yt/frontends/amrvac/api.py
@@ -1,11 +1,7 @@
 """
-API for yt.frontends.amrvac
-
-
-
+frontend API: a submodule that exposes user-facing defs and classes
 """
-
 
 from .data_structures import AMRVACDataset, AMRVACGrid, AMRVACHierarchy
 from .fields import AMRVACFieldInfo
-from .io import AMRVACIOHandler
+from .io import AMRVACIOHandler, read_amrvac_namelist
